fix: add syntax entry points to fix lazy load bug
Fixes #237

The bundler cannot recognize the requires in the stylelint library for
syntaxes, due to the lazy loads in the form `importLazy('package')`
obscuring the require that is taking place.

This patch adds each syntax as its own entry point to bypass this
limitation. It will no longer be necessary when targeting stylelint 14.

see:
https://github.com/stylelint/stylelint/blob/13.13.1/lib/syntaxes/index.js