Fix creating GitHub release during workflow
When I tried the workflow changed in PR #685, the following error occurred with the `git` command:

```
fatal: Failed to resolve 'v1.5.2' as a valid ref.
```

https://github.com/stylelint/vscode-stylelint/actions/runs/15467795091/job/43543973356#step:10:20

To avoid the error, this just removes the use of the `git` command,
since `gh release create` automatically creates a tag if it doesn't exist.
So, the `git` command is no longer needed for this workflow.
Ref https://cli.github.com/manual/gh_release_create