test: improve e2e run script
Ensures that the workspace file is always formatted even if e2e tests
failed, all while allowing for arguments to the test:e2e script to be
passed to vscode-test, which is what is most often desired, rather than
to the prettier cleanup script at the end. Uses the exit code of vscode-
test as the exit code for the e2e script.